Search for deployment script in specified REPO_PATH_DIR
Adapting the approach that is used in Civis's streamlit library here: https://github.com/civisanalytics/civis-services-streamlit/blob/main/entrypoint.sh#L7

This change is intended to make the Docker image more flexible for deployed services within Civis Platform, so that we can use it for arbitrary applications and directories. It still looks for an "start_service" shell script instead of just a Python app script, but that at least can be customized each user's need with this container.